+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/dao/preset/TopicPresetCriteriaMaker.java
@@ -22,6 +22,9 @@ public class TopicPresetCriteriaMaker implements PresetCriteriaMaker {
             case TopicMaintainService.LABEL_LIKE:
                 labelLike(detachedCriteria, objects);
                 break;
+            case TopicMaintainService.ENABLED:
+                enabled(detachedCriteria, objects);
+                break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + s);
         }
@@ -42,6 +45,14 @@ public class TopicPresetCriteriaMaker implements PresetCriteriaMaker {
             String pattern = (String) objects[0];
             detachedCriteria.add(Restrictions.like("label", pattern, MatchMode.ANYWHERE));
             detachedCriteria.addOrder(Order.asc("stringId"));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void enabled(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            detachedCriteria.add(Restrictions.eq("enabled", true));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }